+++ doc/release-notes.md
@@ -13,3 +13,4 @@ This release includes the following features and fixes:
  - Improved help message for backup wallet RPC
  - Various bug fixes that improve node stability and performance
  - Backport getblock RPC's new verbosity mode from bitcoin core for retrieving all transactions of a given block in full.
+ - Added 'parked' state to getchaintips RPC